move get_context
