+++ great_expectations/data_context/__init__.py
@@ -5,4 +5,5 @@ from great_expectations.data_context.data_context import (
     DataContext,
     EphemeralDataContext,
     FileDataContext,
+    get_context,
 )
