+++ great_expectations/data_context/data_context/__init__.py
@@ -7,6 +7,7 @@ from great_expectations.data_context.data_context.base_data_context import (
 from great_expectations.data_context.data_context.cloud_data_context import (
     CloudDataContext,
 )
+from great_expectations.data_context.data_context.context_factory import get_context
 from great_expectations.data_context.data_context.data_context import DataContext
 from great_expectations.data_context.data_context.ephemeral_data_context import (
     EphemeralDataContext,
